Fix search index script. Move lots of lib around.
