+++ lib/node/bus.js
